Move around directories to the standard qiskit form, update measurement to the standard
