+++ qiskit/ignis/characterization/coherence/coherence_utils.py
