+++ qiskit/ignis/characterization/coherence/fitters/basefitter.py
