+++ qiskit/ignis/characterization/coherence/fitters/t1fitter.py
