+++ qiskit/ignis/characterization/coherence/fitters/t2fitter.py
