+++ qiskit/ignis/characterization/coherence/fitters/t2starfitter.py
