+++ qiskit/ignis/randomized_benchmarking/standard_rb/Clifford.py
