+++ qiskit/ignis/randomized_benchmarking/standard_rb/clifford_utils.py
