+++ qiskit/ignis/randomized_benchmarking/standard_rb/rb_fitters.py
