+++ qiskit/ignis/tomography/basis/paulibasis.py
