+++ qiskit/ignis/tomography/fitters/cvx_fit.py
