+++ qiskit/ignis/tomography/fitters/utils.py
